# api-docs.json
{"openapi":"3.1.0","info":{"title":"REST API 서버","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"AdminController","description":"관리자 컨트롤러"},{"name":"UserController","description":"사용자 컨트롤러"},{"name":"ScheduleController","description":"스케줄 컨트롤러"}],"paths":{"/api/calendars/{id}":{"get":{"tags":["calendar-controller"],"operationId":"getCalendarById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["calendar-controller"],"operationId":"updateCalendar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CalendarUpdateDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["calendar-controller"],"operationId":"deleteCalendar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/api/calendars/{calendarId}/schedules/{scheduleId}":{"get":{"tags":["ScheduleController"],"summary":"특정 일정 조회","description":"주어진 일정 ID(scheduleId)를 기반으로 특정 일정을 조회합니다.","operationId":"getSchedule","parameters":[{"name":"calendarId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"scheduleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"500":{"description":"서버 오류","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}},"200":{"description":"일정 조회 성공","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}},"404":{"description":"일정이 존재하지 않음","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["ScheduleController"],"summary":"일정 수정","description":"기존 일정(scheduleId)을 수정합니다.","operationId":"updateSchedule","parameters":[{"name":"calendarId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"scheduleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleRequestDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"일정 수정 성공","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}},"500":{"description":"서버 오류","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}},"404":{"description":"일정이 존재하지 않음","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["ScheduleController"],"summary":"일정 삭제","description":"지정된 일정(scheduleId)을 삭제합니다.","operationId":"deleteSchedule","parameters":[{"name":"calendarId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"scheduleId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"204":{"description":"일정 삭제 성공"},"500":{"description":"서버 오류"},"404":{"description":"일정이 존재하지 않음"}},"security":[{"bearerAuth":[]}]}},"/user":{"post":{"tags":["UserController"],"summary":"내정보 수정","operationId":"modifyUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/modifyUserReqBody"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataVoid"}}}}},"security":[{"bearerAuth":[]}]}},"/api/calendars":{"get":{"tags":["calendar-controller"],"operationId":"getAllCalendars","responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Calendar"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["calendar-controller"],"operationId":"createCalendar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CalendarCreateDto"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{}}}}},"security":[{"bearerAuth":[]}]}},"/api/calendars/{calendarId}/schedules":{"get":{"tags":["ScheduleController"],"summary":"일정 목록 조회","description":"지정된 날짜(date)에 해당하는 일정을 조회합니다.","operationId":"getSchedules","parameters":[{"name":"calendarId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"startDate","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"400":{"description":"잘못된 날짜 형식","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"일정 목록 조회 성공","content":{"application/json;charset=UTF-8":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}}},"500":{"description":"서버 오류","content":{"application/json;charset=UTF-8":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["ScheduleController"],"summary":"일정 생성","description":"주어진 calendarId에 새 일정을 생성합니다.","operationId":"createSchedule","parameters":[{"name":"calendarId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleRequestDto"}}},"required":true},"responses":{"400":{"description":"잘못된 요청 데이터","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"500":{"description":"서버 오류","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}},"200":{"description":"일정 생성 성공","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}}},"security":[{"bearerAuth":[]}]}},"/admin/login":{"post":{"tags":["AdminController"],"summary":"관리자 로그인","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginReqBody"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataLoginDto"}}}}},"security":[{"bearerAuth":[]}]}},"/user/me":{"get":{"tags":["UserController"],"summary":"내 정보","operationId":"me","responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"security":[{"bearerAuth":[]}]}},"/api/calendars/{calendarId}/schedules/daily":{"get":{"tags":["ScheduleController"],"summary":"하루 일정 조회","description":"특정 날짜(24시간)의 일정을 조회합니다.","operationId":"getDailySchedules","parameters":[{"name":"calendarId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"date","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"400":{"description":"잘못된 날짜 형식","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"하루 일정 조회 성공","content":{"application/json;charset=UTF-8":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}}},"500":{"description":"서버 오류","content":{"application/json;charset=UTF-8":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}}}},"security":[{"bearerAuth":[]}]}},"/admin":{"get":{"tags":["AdminController"],"summary":"회원 명단 조회 (페이징, 검색)","operationId":"users","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"searchKeywordType","in":"query","required":false,"schema":{"type":"string","default":"username"}},{"name":"searchKeyword","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataPageDtoUserDto"}}}}},"security":[{"bearerAuth":[]}]}},"/user/{id}":{"delete":{"tags":["UserController"],"summary":"회원 탈퇴 (soft)","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataUserDto"}}}}},"security":[{"bearerAuth":[]}]}},"/admin/logout":{"delete":{"tags":["AdminController"],"summary":"로그아웃","operationId":"logout","responses":{"400":{"description":"Bad Request","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataEmpty"}}}},"200":{"description":"OK","content":{"application/json;charset=UTF-8":{"schema":{"$ref":"#/components/schemas/RsDataVoid"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"Empty":{"type":"object"},"RsDataEmpty":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/Empty"}},"required":["data","msg","resultCode"]},"CalendarUpdateDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"Location":{"type":"object","properties":{"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"address":{"type":"string"}}},"ScheduleRequestDto":{"type":"object","properties":{"title":{"type":"string","maxLength":100,"minLength":1},"description":{"type":"string","maxLength":500,"minLength":0},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"location":{"$ref":"#/components/schemas/Location"},"endTimeValid":{"type":"boolean"}},"required":["endTime","startTime","title"]},"ScheduleResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"calendarId":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"location":{"$ref":"#/components/schemas/Location"},"createDate":{"type":"string","format":"date-time"},"modifyDate":{"type":"string","format":"date-time"}}},"modifyUserReqBody":{"type":"object","properties":{"nickname":{"type":"string"}},"required":["nickname"]},"RsDataVoid":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"type":"object"}},"required":["data","msg","resultCode"]},"CalendarCreateDto":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"}}},"UserLoginReqBody":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"LoginDto":{"type":"object","properties":{"item":{"$ref":"#/components/schemas/UserDto"},"apiKey":{"type":"string"},"accessToken":{"type":"string"}}},"RsDataLoginDto":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/LoginDto"}},"required":["data","msg","resultCode"]},"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"nickname":{"type":"string"},"email":{"type":"string"},"createDate":{"type":"string","format":"date-time"},"modifyDate":{"type":"string","format":"date-time"}}},"Calendar":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"createDate":{"type":"string","format":"date-time"},"modifyDate":{"type":"string","format":"date-time"},"user":{"$ref":"#/components/schemas/SiteUser"},"name":{"type":"string"},"description":{"type":"string"},"sharedUsers":{"type":"array","items":{"$ref":"#/components/schemas/SharedCalendar"}},"messageList":{"type":"array","items":{"$ref":"#/components/schemas/Message"}}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Message":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"createDate":{"type":"string","format":"date-time"},"modifyDate":{"type":"string","format":"date-time"},"user":{"$ref":"#/components/schemas/SiteUser"},"calendar":{"$ref":"#/components/schemas/Calendar"},"content":{"type":"string"},"read":{"type":"boolean"}}},"SharedCalendar":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"calendar":{"$ref":"#/components/schemas/Calendar"},"user":{"$ref":"#/components/schemas/SiteUser"}}},"SiteUser":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"createDate":{"type":"string","format":"date-time"},"modifyDate":{"type":"string","format":"date-time"},"username":{"type":"string"},"nickname":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["ADMIN","USER"]},"apiKey":{"type":"string"},"deletedDate":{"type":"string","format":"date-time"},"deleted":{"type":"boolean"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}}}},"PageDtoUserDto":{"type":"object","properties":{"currentPageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"totalPages":{"type":"integer","format":"int64"},"totalItems":{"type":"integer","format":"int64"},"items":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}},"required":["currentPageNumber","items","pageSize","totalItems","totalPages"]},"RsDataPageDtoUserDto":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/PageDtoUserDto"}},"required":["data","msg","resultCode"]},"RsDataUserDto":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/UserDto"}},"required":["data","msg","resultCode"]}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}